backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
